babysitter first test

+++ src/main/java/com/babysitterapp/Babysitter.java
@@ -4,7 +4,7 @@ public class Babysitter {
 
     // Bedtime is 9 PM
 
-    public int calculateRate(int startTime, int endTime) {
+    public int calculateRate(double startTime, double endTime) {
         int rate = 0;
         return rate;
     }
